Added charges, fixed cleanup.py (cleanup_windows.bat), fixed coat of arms generation
Still bugs and glaring issues with coat of arms generation. In particular, rule of tinctures along with other coat of arms generation is not strictly followed or accounted for. Current version has a few bugs and poor generations. Will need refinement, further development, and modification in future versions to improve overall coat of arms generation. For now, the basic program is able to run, generate basic (yet frequently distasteful) coat of arms, and has basic functionality with regards to bat files, directory structure, and file importing.